requirements.txt
pandas
matplotlib
numpy
scikit-learn
tensorflow
keras
seaborn
ipykernel
python-dotenv
psycopg2-binary 